+++ data/src/main/java/com/susu/data/remote/model/response/FriendResponse.kt
@@ -3,4 +3,4 @@ package com.susu.data.remote.model.response
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class FriendResponse(val id: Int)
+data class FriendResponse(val id: Long)
